+++ src/index.d.ts
@@ -102,6 +102,7 @@ import {
 	deletePost,
 	fetchCommunityGuidelines,
 	fetchPosts,
+	jumpToPost,
 	likePost,
 	search,
 	unlikePost,
@@ -251,6 +252,7 @@ import {
 
 import {
 	confirmEmailChange,
+	deleteAccount,
 	requestEmailChange,
 	resetPassword,
 	sendAccountSetupEmail,
@@ -277,7 +279,10 @@ import {
 	blockedUsers,
 	deletePicture,
 	getUserData,
-	isDisplayNameAvailable,
+	getUserSignature,
+	isUsernameAvailable,
+	setUserSignature,
+	toggleSignaturePrivate,
 	unblockUser,
 	updateDisplayName,
 	uploadPicture,
@@ -287,8 +292,9 @@ import {
 import {
 	fetchMemberships,
 	fetchRechargeTokens,
+	restorePurchases,
 	upgradeSubscription
-} from './services/user/memberships.js';
+} from './services/user/memberships.ts';
 
 import {
 	deleteNotification,
@@ -304,6 +310,10 @@ import {
 	startLiveEventPolling,
 	updateNotificationSetting
 } from './services/user/notifications.js';
+
+import {
+	startOnboarding
+} from './services/user/onboarding.ts';
 
 import {
 	fetchCustomerPayments
@@ -377,6 +387,7 @@ declare module 'musora-content-services' {
 		createPost,
 		createPracticeNotes,
 		createThread,
+		deleteAccount,
 		deleteComment,
 		deleteItemsFromPlaylist,
 		deleteNotification,
@@ -523,6 +534,7 @@ declare module 'musora-content-services' {
 		getUserData,
 		getUserMonthlyStats,
 		getUserPractices,
+		getUserSignature,
 		getUserWeeklyStats,
 		getWeekNumber,
 		globalConfig,
@@ -531,10 +543,11 @@ declare module 'musora-content-services' {
 		isBucketUrl,
 		isContentLiked,
 		isContentLikedByIds,
-		isDisplayNameAvailable,
 		isNextDay,
 		isSameDate,
+		isUsernameAvailable,
 		jumpToContinueContent,
+		jumpToPost,
 		likeComment,
 		likeContent,
 		likePlaylist,
@@ -580,17 +593,21 @@ declare module 'musora-content-services' {
 		restoreItemFromPlaylist,
 		restoreNotification,
 		restorePracticeSession,
+		restorePurchases,
 		restoreUserActivity,
 		restoreUserPractice,
 		search,
 		sendAccountSetupEmail,
 		sendPasswordResetEmail,
 		setStudentViewForUser,
+		setUserSignature,
 		setupAccount,
 		startLiveEventPolling,
+		startOnboarding,
 		status,
 		toDayjs,
 		togglePlaylistPrivate,
+		toggleSignaturePrivate,
 		unEnrollUserInGuidedCourse,
 		unassignModeratorToComment,
 		unblockUser,